# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.Error:
    properties:
      code:
        type: integer
      errors:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
  domain.Success:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  domain.UserStoreRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role_id:
        type: integer
    required:
    - email
    - name
    - password
    - role_id
    type: object
  domain.UserUpdateRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role_id:
        type: integer
    type: object
...
host: localhost:3000
info:
  contact: {}
  title: Book Store API Documentation
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Get list of users
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Size of page (default 10)
        in: query
        name: size
        type: integer
      - description: Search query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/domain.Success'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Get list of user
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Store user
      parameters:
      - description: user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.UserStoreRequest'
      produces:
      - application/json
      responses:
        "201":
          description: user detail
          schema:
            $ref: '#/definitions/domain.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      security:
      - Bearer: []
      summary: Store user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete User
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success delete user
          schema:
            $ref: '#/definitions/domain.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      security:
      - Bearer: []
      summary: Delete User
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get user by id
      parameters:
      - description: user ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: user detail
          schema:
            $ref: '#/definitions/domain.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Get user by id
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: user ID
        in: path
        name: id
        required: true
        type: integer
      - description: user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: user detail
          schema:
            $ref: '#/definitions/domain.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      security:
      - Bearer: []
      summary: Update user
      tags:
      - users
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"